fix(stella-graph): drop the child_count cast the bump made redundant
tree-sitter 0.27.0 widened `Node::child_count` from `usize` to `u32`,
so the six `(0..node.child_count() as u32)` loops now cast u32 to u32.
Clippy's `unnecessary_cast` fires on each, and `windows-check` runs
with `-D warnings`, so all six were errors there — `cargo check` and
`msrv` pass, which is why this survived the first fix on this branch.

`child(i: u32)` was widened in the same release, so the loop variable
still matches the accessor it feeds and no other conversion moves.

Verified: `cargo clippy -p stella-graph --all-targets --locked
-- -D warnings` exit 0, `cargo test -p stella-graph --locked` 297
passed / 0 failed, `cargo fmt --check` clean.

+++ crates/stella-graph/src/parse.rs
@@ -558,7 +558,7 @@ fn extract_rust_orm_tables(root: Node, src: &[u8]) -> Vec<Symbol> {
             }
         }
 
-        for idx in (0..node.child_count() as u32).rev() {
+        for idx in (0..node.child_count()).rev() {
             if let Some(child) = node.child(idx) {
                 stack.push(child);
             }
@@ -594,7 +594,7 @@ fn extract_python_orm_tables(root: Node, src: &[u8]) -> Vec<Symbol> {
             }
         }
 
-        for idx in (0..node.child_count() as u32).rev() {
+        for idx in (0..node.child_count()).rev() {
             if let Some(child) = node.child(idx) {
                 stack.push(child);
             }
